bin/requirements.txt
boto3
click
python-graphql-client
pyyaml
webflowpy
docker_registry_client_async
asyncio